Moved database method from Operations class to new Database class
